% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-package.R
\name{compile_package}
\alias{compile_package}
\title{Compile all \code{quick()} functions in a package.}
\usage{
compile_package(path = ".")
}
\arguments{
\item{path}{Path to an R package}
}
\value{
Called for it's side effect.
}
\description{
This will compile all \code{quick()} functions in an R package, and
generate source files in the \verb{src/} directory.
}
\details{
Note, this function is automatically invoked during a \code{pkgload::load_all()} call.
}
