% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.R
\name{quick}
\alias{quick}
\title{Quick Function}
\usage{
quick(fun, name = NULL)
}
\arguments{
\item{fun}{An R function}

\item{name}{Optional string, name to use for the function.}
}
\value{
A quicker R function.
}
\description{
Quick Function
}
